Create Content comopnent for use in other components
